fix(service_discovery): a record detection for certain devices
Ensure that the A record for certain devices (for example shellies)
is done correctly by buffering currently irrelevant A records as
they get sent before the SRV record for the specific target is sent.

This change slightly increases memory overhead as an additional
configurable buffer needs to be added to allow buffering A
records which are not yet relevant for an SRV response.