Usability adjustments to the Pico docker files
These adjustments shift the compilation action in the docker container from build-time to run-time. This allows the user to mount the build directory on their local system to ease retrieval of the built artifacts. Additionally, I've included some documentation for the relevant paths where the artifacts are created so it isn't necessary to instantiate a docker container to browse the filesystem looking for the firmware to then use "docker cp" to retrieve it from the container.