fix(mdns_client): support for rp2 on 1.24
As reported in https://github.com/cbrand/micropython-mdns/pull/30 there were issues even after the patches wtih the 1.24 version of resolving MDNS. This was due to the socket binding not being bound to the MDNS address. Both on the ESP32 and the RP2 with 1.24 binding works now again.

To not have a backwards issue with the code base, dropping support for anything older than 1.24 in this commit and in the next release.